pyproject.toml
[tool.poetry]
name = "cours_work_5"
version = "0.1.0"
description = ""
authors = ["sergey <szaprudnov@hotmail.com>"]
readme = "README.md"
packages = [{include = "cours_work_5"}]

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.29.0"
psycopg2-binary = "^2.9.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"